requirements.txt
Flask
python-socketio==4.2.0
python-engineio==3.8.2
eventlet
psycopg2-binary
redis
flask-cors
PyJWT